// src/pages/transaction/ordering/enums.ts
export enum PaymentMethod {
    CASH = 'CASH',
    CHECK = 'CHECK',
    DEBIT_CARD = 'DEBIT CARD',
    CREDIT_CARD = 'CREDIT CARD',
    GIFT_CHECK = 'GIFT CHECK',
    OTHER_PAYMENT = 'OTHER PAYMENT',
    FREE = 'FREE'
}

export enum PaymentStatus {
    START = 'STARTED',
    ONGOING = 'ONGOING',
    COMPLETED = 'COMPLETED'
}